requirements.txt
nose==1.3.4
flake8==2.2.3
coverage==3.7.1
mock==1.0.1
PyYAML==3.11
